NewKmc/requirements.txt
2to3==1.0
appdirs==1.4.4
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
atlassian==0.0.0
atlassian-python-api==3.25.0
atomicwrites==1.4.1
attrs==22.1.0
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
black==22.6.0
boto3==1.24.39
botocore==1.27.39
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
comtypes==1.1.11
configparser==5.2.0
crayons==0.4.0
cryptography==37.0.4
cssselect==1.1.0
cssutils==2.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distro-info==1.0
Django==4.0.6
exceptiongroup==1.0.4
executing==0.9.1
fasteners==0.17.3
fastjsonschema==2.16.1
filelock==3.7.1
Flask==2.1.3
h11==0.13.0
idna==3.3
image==1.5.33
importlib-metadata==4.12.0
iniconfig==1.1.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
jeepney==0.8.0
Jinja2==3.1.2
jira==3.3.1
jmespath==1.0.1
jsonschema==4.8.0
jupyter-core==4.11.1
KalturaApiClient==18.17.0
keyring==23.7.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
MouseInfo==0.1.3
MultipartPostHandler==0.1.0
MultipartPostHandler2==0.1.5
mypy-extensions==0.4.3
mysql-connector==2.2.9
nbformat==5.4.0
networkx==2.8.5
numpy==1.23.1
oauthlib==3.2.0
onesecmail==0.0.1
outcome==1.2.0
packaging==21.3
paramiko==2.11.0
parso==0.8.3
pathspec==0.9.0
pbr==5.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip-review==1.1.1
platformdirs==2.5.2
pluggy==1.0.0
poster3==0.8.1
premailer==3.10.0
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
PyAutoGUI==0.9.53
PyAutoIt==0.6.5
pycparser==2.21
pydot==1.4.2
pydotplus==2.0.2
PyGetWindow==0.0.9
Pygments==2.12.0
PyJWT==2.4.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.1
pywin32==305
pywin32-ctypes==0.2.0
pywinauto==0.6.8
PyYAML==6.0
pyzbar==0.1.9
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
retrying==1.3.3
s3transfer==0.6.0
SecretStorage==3.3.2
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sos==0.23.4
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
tabulate==0.9.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==4.3.0
tzdata==2022.1
uritools==4.0.0
urlextract==1.6.0
urllib3==1.26.11
wcwidth==0.2.5
webdriver-manager==3.8.2
webdrivermanager==0.10.0
Werkzeug==2.2.1
wget==3.2
wrapt==1.14.1
wsproto==1.1.0
xlrd==2.0.1
xxhash==3.0.0
yagmail==0.15.280
zipp==3.8.1
pytest-dotenv